Memory update

+++ source/Memory.cpp
@@ -7,13 +7,13 @@ Memory::Memory(int size) {
 
 void Memory::Clear() {
 	for (unsigned int i = 0; i < memory.size(); i++) {
-		memory[(int)i] = 0;
+		memory[i] = 0;
 	}
 }
 
 bool Memory::Empty() {
 	for (unsigned int i = 0; i < memory.size(); i++) {
-		if (memory[(int)i] != 0) {
+		if (memory[i] != 0) {
 			return false;
 		}
 	}
